Reimplemented markMask method of QuadTree.
markMask is now generateMask. It works by forwarding
points from the current node level to its child nodes;
if the density of points (pts/pixel) is larger than a threshold,
the node region is filled with black, else, it is filled with white.
The algorithm proceeds recursively marking smaller rectangular regions
with white in which keypoints should be detected within.