Changed ReconstructionVisualizer::addKeyFrame
The function no longer needs the second parameter
informing the keyframe text (which was redundant).
This information is now set as the string "kf"
concatenated with the keyframe index.

The text member variable was also removed from
the Keyframe struct (in common.h), since it
was not being used anyway.