Added the time_stamp parameter to the writePoseToFile method.
For the OpticalFlowVisualOdometry class, the writeToPoseFile method was added to save the calculated pose with the time stamp already included.

In addition, the RGBDLoader class was modified to return the timestamp of the RGB and depth images. Thus, through the methods getNextRgbImageTimeStamp and getNextDepthImageTimeStamp it is possible to obtain the time stamps corresponding to these images.